encapsulate VTK in a separate file
